Updated app.py, queue.html
The functionality I updated in app.py is if we re-upload the incomplete_records csv file when we comeback again the completed records will not appear in the queue. In queue.html file there was a functionality to show the record is locked. Now, the record itself will nor appear in the queue. There is no need of this functionality. So, I removed it.